requirements.txt
requests
beautifulsoup4
pandas
sqlalchemy
psycopg2-binary
python-dotenv
dbt-core
dbt-postgres
pymysql
cryptography